Move facet button content to their own files
